conan requirement removed

+++ src/util/Utils.h
@@ -12,9 +12,9 @@ class Utils {
 public:
     static std::vector<std::vector<int>> loadGEucSpace(const std::string &file_path);
 
-    static std::vector<std::vector<int>> readVList(const std::string &file_path);
+    static std::vector<std::vector<float>> readVList(const std::string &file_path);
 
-    static float stdDev(std::vector<float> &items, float average);
+    static float stdDev(std::vector<int> &items, float average);
 
     static void floydWarshall(std::vector<std::vector<float>>& G);
